fix food_formula/serializers -> imports GOOD | some fix food_formula/views

+++ backend/food_formula/views.py
@@ -56,13 +56,12 @@ class RecipeViewSet(viewsets.ModelViewSet):
                 recipe, context={'request': request}
             )
             return Response(serializer.data, status=status.HTTP_201_CREATED)
-        elif request.method == "DELETE":
-            if recipe_in_cart.exists():
-                recipe_in_cart.delete()
-                return Response(
-                    {'msg': 'Успешно удалено'},
-                    status=status.HTTP_204_NO_CONTENT
-                )
+        elif request.method == "DELETE" and recipe_in_cart.exists():
+            recipe_in_cart.delete()
+            return Response(
+                {'msg': 'Успешно удалено'},
+                status=status.HTTP_204_NO_CONTENT
+            )
         return Response(
             {'error': errors['recipe_not_in']},
             status=status.HTTP_400_BAD_REQUEST
